проверка на существование файла

+++ src/files/files.service.ts
@@ -29,7 +29,7 @@ export class FilesService {
     async deleteFile(fileName) {
         try {
             const filePath = path.resolve(__dirname, '..', `static/${fileName}`)
-            fs.unlinkSync(filePath)
+            if (fs.existsSync(filePath)) fs.unlinkSync(filePath)
         } catch (e) {
             throw new HttpException('Произошла ошибка при удалении файла', HttpStatus.INTERNAL_SERVER_ERROR)
         }
